refactor(openomni): recorded verifier input moves to evidence (#606)
The completion-admission split scoping ruled NO-GO on a semantic
split (two production consumers both pull across every section; the
file is one transaction protocol whose re-assertion of head/basis/
lease at every step IS the product; the seams a split would need are
the exact role-noun suffixes AGENTS.md bans). One extraction is
earned: durableVerifierInput + createDurableCompletionResultAuthority-
Port evolve with the VerifierRegistry obligation/read-back vocabulary,
not the admission protocol, and have a second consumer (worker-
completion read-back). They move to evidence/verifier-recorded-input,
taking the VerifierRegistry and zod imports with them; the admission
port contract stays where it is via a type-only back-edge (0 value
cycles). Pure move — bodies verbatim, consumers repointed, no
re-export residue.
